+++ lib/danger/danger_core/plugins/dangerfile_messaging_plugin.rb
@@ -8,12 +8,13 @@ module Danger
   #
   # The message within which Danger communicates back is amended on each run in a session.
   #
-  # Each of `message`, `warn` and `fail` have a `sticky` flag, `true` by default, which
-  # means that the message will be crossed out instead of being removed. If it's not use on
-  # subsequent runs.
+  # Each of `message`, `warn` and `fail` have a `sticky` flag, `false` by default, which
+  # when `true` means that the message will be crossed out instead of being removed.
+  # If it's not called again on subsequent runs.
   #
   # By default, using `fail` would fail the corresponding build. Either via an API call, or
-  # via the return value for the danger command.
+  # via the return value for the danger command. If you have linters with errors for this call
+  # you can use `messaging.fail` instead.
   #
   # It is possible to have Danger ignore specific warnings or errors by writing `Danger: Ignore "[warning/error text]"`.
   #
@@ -94,7 +95,7 @@ module Danger
     #           Optional. The line in the file to present the message in.
     # @return   [void]
     #
-    def message(message, sticky: true, file: nil, line: nil)
+    def message(message, sticky: false, file: nil, line: nil)
       @messages << Violation.new(message, sticky, file, line)
     end
 
@@ -112,7 +113,7 @@ module Danger
     #           Optional. The line in the file to present the message in.
     # @return   [void]
     #
-    def warn(message, sticky: true, file: nil, line: nil)
+    def warn(message, sticky: false, file: nil, line: nil)
       return if should_ignore_violation(message)
       @warnings << Violation.new(message, sticky, file, line)
     end
@@ -131,7 +132,7 @@ module Danger
     #           Optional. The line in the file to present the message in.
     # @return   [void]
     #
-    def fail(message, sticky: true, file: nil, line: nil)
+    def fail(message, sticky: false, file: nil, line: nil)
       return if should_ignore_violation(message)
       @errors << Violation.new(message, sticky, file, line)
     end