+++ CHANGELOG.md
@@ -271,6 +271,7 @@
 * Fix Buildkite repo slug URL generation - phillfarrugia
 * Added support for [Bitrise](http://bitrise.io) - kylejm
 * Set LANG variable to en_US.UTF-8 before running git commands - bogren
+* Switch messaging's sticky to default to `false` - Bogren
 
 ## 3.0.3